+++ src/main/java/sg/edu/iss/team6/repository/StudentRepository.java
@@ -1,6 +1,7 @@
 package sg.edu.iss.team6.repository;
 
 import java.util.ArrayList;
+import java.util.function.LongBinaryOperator;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -9,7 +10,7 @@ import sg.edu.iss.team6.model.Student;
 import sg.edu.iss.team6.model.User;
 
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
 
     Student findByStudentId(Long studentId);
     Student findByUserUsername(User u);
